add basic room tables and screens

+++ app/src/main/java/com/sherryyuan/wordy/database/EntryDao.kt
@@ -18,7 +18,7 @@ interface EntryDao {
     suspend fun insertEntry(entry: Entry)
 
     @Query(value = "UPDATE Entry SET timestamp= :timestamp, wordCount= :wordCount, projectId = :projectId WHERE id = :entryId")
-    suspend fun updateEntry(entryId: Int, timestamp: Long, wordCount: Int, projectId: Int)
+    suspend fun updateEntry(entryId: Long, timestamp: Long, wordCount: Int, projectId: Int)
 
     @Delete
     suspend fun deleteEntry(entry: Entry)
